Add Preliminary Bubble
This change adds a preliminary bubble above the timeline. The
positioning isn't quite worked out yet. Currently it has some hardcoded
values.

There are also lines attaching the bubble to the timeline. The start to
finish calculation works, but there is more positioning problems when
it comes to getting the coordinates of the bubble corners. I think it
has something to do with the CSS position.