Draw bubbles with respect to the years on the timeline
The events parameter drives the creation of bubbles. The bubbles are lazily
drawn on the page with a hardcoded left shift (50px) from the target date. The
lines point to the correct target date. Only if the date is off the left side
off the screen will the bubble not be drawn.